Improved string support
- Got rid of bak files (back up files)
- Added possibility for adding multiple strings on 1 line
(this would before cause unexpected behavior)
- Fixed issues with strings not completely acting as defined, more
specifically when using octal or hexadecimal representations of single
characters in a string, before it would expect exactly 3 octal or
hexadecimal digits, now it acts like in C
- Got rid of 2 redundant files "ConstantCodePart.java" and
"StringCodePart.java" as they are now merged into "LiteralCodePart.java"
- Fixed having a single label as constant code part

WARNING:
The changes are not yet thoroughly tested and might contain some
glitches. When these glitches get noticed, they will get fixed in future
commits.

WARNING:
An uncaught null pointer exception breaking the program was noticed
during a test run.
The stack trace contained no link to source code.
It was not reproduceable.
The odds are that the new modifications in this commit have nothing to
do with the issue.